Broke out tasks into dep and history.
