+++ Jakefile.js
@@ -4,7 +4,6 @@
 
 var fs = require("fs");
 var child_process = require("child_process");
-
 var Ronn = require("ronn").Ronn;
 var rimraf = require("rimraf");
 var walk = require("walk");
@@ -78,7 +77,11 @@ task("default", ["install"]);
 
 desc("Install all modules including devDependencies");
 task("install", function () {
-    spawn("npm", ["install"], complete);
+    var dep = jake.Task['dep'];
+    dep.addListener('complete', function () {
+        spawn("npm", ["install"], complete);
+    });
+    dep.invoke();
 }, {
     async: true
 });
@@ -149,4 +152,18 @@ desc("Remove development tools");
 task("maintainer-clean", function () {
     spawn("rpm", ["rm", "webkit-devtools-agent"]);
     nuke("tools");
+});
+
+desc("Fetch external dependencies");
+task("dep", function () {
+    spawn(process.argv[0], ["./scripts/fetch_deps.js"], complete);
+}, {
+    async: true
+});
+
+desc("Print history");
+task("history", function () {
+    spawn(process.argv[0], ["./scripts/postinstall.js"], complete);
+}, {
+    async: true
 });
